LayoutTests/http/tests/inspector/resource-tree/appcache-iframe-manifests-expected.txt

Tests that application cache model keeps track of manifest urls and statuses correctly.

Bug 64581
Dumping application cache tree:
    (empty)
Dumping application cache model:
    (empty)

Dumping application cache tree:
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1
        Frame: frame1 (page-with-manifest.php)
Dumping application cache model:
    Frame: frame1 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=1) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1

Dumping application cache tree:
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
        Frame: frame1 (page-with-manifest.php)
Dumping application cache model:
    Frame: frame1 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=2) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2

Dumping application cache tree:
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
        Frame: frame1 (page-with-manifest.php)
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1
        Frame: frame2 (page-with-manifest.php)
Dumping application cache model:
    Frame: frame1 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=2) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
    Frame: frame2 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=1) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1

Dumping application cache tree:
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
        Frame: frame1 (page-with-manifest.php)
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1
        Frame: frame2 (page-with-manifest.php)
        Frame: frame3 (page-with-manifest.php)
Dumping application cache model:
    Frame: frame1 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=2) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
    Frame: frame2 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=1) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1
    Frame: frame3 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=1) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1

Dumping application cache tree:
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
        Frame: frame1 (page-with-manifest.php)
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1
        Frame: frame2 (page-with-manifest.php)
Dumping application cache model:
    Frame: frame1 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=2) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
    Frame: frame2 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=1) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=1

Dumping application cache tree:
    Manifest URL: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2
        Frame: frame1 (page-with-manifest.php)
Dumping application cache model:
    Frame: frame1 (http://127.0.0.1:8000/inspector/resource-tree/resources/page-with-manifest.php?manifestId=2) has manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=2

Dumping application cache tree:
    (empty)
Dumping application cache model:
    (empty)
